Reorganize the curation apis
