+++ backend/corpora/lambdas/api/v1/curation/collections/collection_uuid/dataset.py
@@ -1,17 +1,34 @@
 import json
-import boto3
 import logging
-from flask import g, make_response, jsonify, request
+
+import boto3
+from flask import g, request, make_response, jsonify
 
 from backend.corpora.api_server.db import dbconnect
 from backend.corpora.common.corpora_config import CorporaConfig
 from backend.corpora.common.corpora_orm import CollectionVisibility
-from backend.corpora.lambdas.api.v1.common import get_collection
+from backend.corpora.common.entities import Dataset
+from backend.corpora.common.utils.http_exceptions import InvalidParametersHTTPException
 from backend.corpora.lambdas.api.v1.authorization import owner_or_allowed
+from backend.corpora.lambdas.api.v1.common import get_collection
+from backend.corpora.lambdas.api.v1.dataset import delete_dataset_common
+
+
+@dbconnect
+def delete_dataset(token_info: dict, collection_uuid: str, curator_tag: str = None, dataset_uuid=None):
+    db_session = g.db_session
+    if dataset_uuid:
+        dataset = Dataset.get(db_session, dataset_uuid, include_tombstones=True)
+    elif curator_tag:
+        dataset = Dataset.get_dataset_from_curator_tag(db_session, collection_uuid, curator_tag)
+    else:
+        raise InvalidParametersHTTPException()
+    delete_dataset_common(db_session, dataset, token_info)
+    return "", 202
+
 
 sts_client = boto3.client("sts")
 logger = logging.getLogger(__name__)
-
 duration = 43200
